pyproject.toml
[project]
name = "capedge-client"
version = "0.1.0"
description = "Python client for the CapEdge API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["capedge_client"]
